Top menu changes
I have removed the dropdown menu for the checkout, instead this will be a direct link to the checkout page.  

I have also changed the search button (magnifying glass) into a custom button (based upon Bootstrap buttons).

I have also added a class of home to the home item in the main navigation, this is to help with styling.

I have also changed the button in the full page search bar to a larger version of the custom button mentioned above.